+++ scripts/publish-helpers/update-package-json-files.ts
@@ -12,6 +12,7 @@ const manualPackagesToFix = [
   "@macrostrat/style-system",
   "@macrostrat/web-components-bundler",
   "@macrostrat/hyperstyle-loader",
+  "@macrostrat/vite-plugin-hyperstyles",
 ];
 
 const packageJSONKeyOrder = [